refactor: consolidate SWML verb handlers into single SWMLHandler.ts
Mirror the Python SDK's single-file layout
(signalwire/signalwire/core/swml_handler.py) by merging
SWMLVerbHandler.ts, AIVerbHandler.ts, and VerbHandlerRegistry.ts
into src/SWMLHandler.ts. Public API surface unchanged - same
exported names, same types, same behavior. Internal cross-file
imports between the three files become in-file references.
